+++ gateway-service-impl/src/main/java/org/hypertrace/gateway/service/entity/query/ExecutionTreeBuilder.java
@@ -15,7 +15,6 @@ import org.hypertrace.core.attribute.service.v1.AttributeMetadata;
 import org.hypertrace.core.attribute.service.v1.AttributeSource;
 import org.hypertrace.gateway.service.common.util.ExpressionReader;
 import org.hypertrace.gateway.service.common.util.TimeRangeFilterUtil;
-import org.hypertrace.gateway.service.entity.config.TimestampConfigs;
 import org.hypertrace.gateway.service.entity.query.visitor.ExecutionContextBuilderVisitor;
 import org.hypertrace.gateway.service.entity.query.visitor.FilterOptimizingVisitor;
 import org.hypertrace.gateway.service.entity.query.visitor.PrintVisitor;
@@ -45,7 +44,7 @@ public class ExecutionTreeBuilder {
                 executionContext.getEntitiesRequest().getEntityType());
 
     this.sourceSetsIfFilterAndOrderByAreFromSameSourceSets =
-        ExecutionTreeUtils.getSourceSetsIfFilterAndOrderByAreFromSameSourceSets(
+        ExpressionReader.getSourceSetsIfFilterAndOrderByAreFromSameSourceSets(
             executionContext.getExpressionContext());
   }
 
@@ -71,7 +70,7 @@ public class ExecutionTreeBuilder {
     // (live + non live) does not make sense, since filters on any other data source will anyways
     // filter out the "non live" entities
     boolean areFiltersOnlyOnEds =
-        ExecutionTreeUtils.areFiltersOnlyOnCurrentDataSource(
+        ExpressionReader.areFiltersOnlyOnCurrentDataSource(
             executionContext.getExpressionContext(), EDS.name());
     if (entitiesRequest.getIncludeNonLiveEntities() && areFiltersOnlyOnEds) {
       ExecutionTreeUtils.removeDuplicateSelectionAttributes(executionContext, EDS.name());
@@ -106,7 +105,8 @@ public class ExecutionTreeBuilder {
     // can be source specification
     // optimization where all projections, filters, order by, sort and limit can be pushed down to
     // the data store
-    Optional<String> singleSourceForAllAttributes = getValidSingleSource();
+    Optional<String> singleSourceForAllAttributes =
+        ExecutionTreeUtils.getValidSingleSource(executionContext);
     if (singleSourceForAllAttributes.isPresent()) {
       String source = singleSourceForAllAttributes.get();
       QueryNode rootNode = buildExecutionTreeForSameSourceFilterAndSelection(source);
@@ -364,55 +364,5 @@ public class ExecutionTreeBuilder {
 
   private QueryNode createPaginateOnlyNode(QueryNode queryNode, EntitiesRequest entitiesRequest) {
     return new PaginateOnlyNode(queryNode, entitiesRequest.getLimit(), entitiesRequest.getOffset());
-  }
-
-  private Optional<String> getValidSingleSource() {
-    EntitiesRequest entitiesRequest = executionContext.getEntitiesRequest();
-
-    Optional<String> singleSourceForAllAttributes =
-        ExecutionTreeUtils.getSingleSourceForAllAttributes(executionContext.getExpressionContext());
-
-    if (singleSourceForAllAttributes.isEmpty()) {
-      return Optional.empty();
-    }
-
-    /**
-     * Entities queries usually have an inherent time filter, via {@link
-     * EntitiesRequest#getStartTimeMillis()} and {@link EntitiesRequest#getEndTimeMillis()}.
-     *
-     * <p>The time filter is usually applied on QS, apart from few entity types, which have a
-     * timestamp column specified through timestamp config {@link TimestampConfigs}
-     *
-     * <p>Single source for all attributes doesn't take care of the inherent time filter for the
-     * single source
-     *
-     * <p>If the single source is {@link EDS} and the timestamp column doesn't exist for the
-     * corresponding entity type, the query will be executed on {@link EDS} source, without the
-     * inherent filter, leading to erroneous outputs
-     *
-     * <p>Hence, default to non single source, if
-     *
-     * <ul>
-     *   <li>single source is {@link EDS}
-     *   <li>inherent time filter is valid
-     *   <li>only live entities are requested (since non live entities request ignores the inherent
-     *       time filter anyways)
-     *   <li>timestamp column doesn't exist for the corresponding entity type
-     * </ul>
-     */
-    String singleSource = singleSourceForAllAttributes.get();
-    if (EDS.name().equals(singleSource)
-        && isValidTimeRange(
-            entitiesRequest.getStartTimeMillis(), entitiesRequest.getEndTimeMillis())
-        && !entitiesRequest.getIncludeNonLiveEntities()
-        && TimestampConfigs.getTimestampColumn(entitiesRequest.getEntityType()) == null) {
-      return Optional.empty();
-    }
-
-    return singleSourceForAllAttributes;
-  }
-
-  private static boolean isValidTimeRange(long startTimeMillis, long endTimeMillis) {
-    return startTimeMillis != 0 && endTimeMillis != 0 && startTimeMillis < endTimeMillis;
   }
 }
