+++ gateway-service-impl/src/main/java/org/hypertrace/gateway/service/explore/ExploreService.java
@@ -18,8 +18,8 @@ import org.hypertrace.entity.v1.entitytype.EntityType;
 import org.hypertrace.gateway.service.common.AttributeMetadataProvider;
 import org.hypertrace.gateway.service.common.ExpressionContext;
 import org.hypertrace.gateway.service.common.config.ScopeFilterConfigs;
+import org.hypertrace.gateway.service.common.util.ExpressionReader;
 import org.hypertrace.gateway.service.entity.config.EntityIdColumnsConfigs;
-import org.hypertrace.gateway.service.entity.query.ExecutionTreeUtils;
 import org.hypertrace.gateway.service.explore.entity.EntityRequestHandler;
 import org.hypertrace.gateway.service.v1.explore.ExploreRequest;
 import org.hypertrace.gateway.service.v1.explore.ExploreResponse;
@@ -128,8 +128,7 @@ public class ExploreService {
               request.getTimeAggregationList(),
               request.getOrderByList(),
               request.getGroupByList());
-      Optional<String> source =
-          ExecutionTreeUtils.getSingleSourceForAllAttributes(expressionContext);
+      Optional<String> source = ExpressionReader.getSingleSourceForAllAttributes(expressionContext);
       if (source.isPresent() && EDS.toString().equals(source.get())) {
         return entityRequestHandler;
       }
